Add dynamic registration mechanism for adding new generators (initial commit)
